refactor: move duplicated code to a helper file
