+++ src/presentation/test/index.ts
@@ -1,3 +1,4 @@
+export * as Helper from "./form-helper"
 export * from "./mock-authentication";
 export * from "./mock-validation";
 export * from "./mock-save-access-token";
